Release 11 with simplified serial port management
The serial port is now managed in the main program. No need to dig into
the proxySerial library.
The serial port could be: hardware, software or I2C.
Also added, speed management.
Beware: older example except port_main may need some work for release
11 compatibility.

+++ i2cSerial.cpp
@@ -12,7 +12,8 @@
 
 #include "WProgram.h"
 #include <Wire.h>
-#include "i2cSerial.h";
+#include "i2cSerial.h"
+#include "Stream.h"
 
 // ---------------- Functions
 
@@ -77,7 +78,7 @@ String i2cSerial::WhoAmI() {
 }
 
 
-void i2cSerial::begin(uint16_t b) {
+void i2cSerial::begin(long b) {
 
   _writeTo(_address, 0x0e << 3, 0x01 << 3); // software reset
